requirements.txt
appdirs==1.4.3
asgiref==3.2.10
backcall==0.2.0
CacheControl==0.12.6
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
decorator==4.4.2
defusedxml==0.7.0rc1
distlib==0.3.0
distro==1.4.0
Django==3.1
django-appconf==1.0.4
django-environ==0.4.5
django-extra-fields==3.0.0
django-filter==2.3.0
django-imagekit==4.0.2
django-phone-field==1.8.1
django-templated-mail==1.1.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.5.0
djoser==2.0.5
docopt==0.6.2
drf-writable-nested==0.6.1
environ==1.0
greenlet==0.4.16
html5lib==1.0.1
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.1
install==1.3.4
ipaddr==2.2.0
ipdb==0.13.3
ipython==7.16.1
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.2
lockfile==0.12.2
Markdown==3.2.2
MarkupSafe==1.1.1
msgpack==1.0.0
newsapi-python==0.2.6
oauthlib==3.1.0
packaging==20.3
parso==0.7.1
pep517==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pilkit==2.0
Pillow==7.2.0
progress==1.5
prompt-toolkit==3.0.7
psycopg2==2.8.5
ptyprocess==0.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.6.1
PyJWT==1.7.1
pynvim==0.4.1
pyparsing==2.4.6
python3-openid==3.2.0
pytoml==0.1.21
pytz==2020.1
pyvim==3.0.2
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
zipp==3.1.0
